Issue #SB-14499 fix: Assessment metrics code refactor
